updating and merging
